+++ src/main/java/org/folio/entlinks/config/constants/ErrorCode.java
@@ -3,6 +3,7 @@ package org.folio.entlinks.config.constants;
 import lombok.Getter;
 
 public enum ErrorCode {
+  UNKNOWN_CONSTRAINT("-1", "Unknown constraint."),
   NO_SUGGESTIONS("101", ""),
   MORE_THAN_ONE_SUGGESTIONS("102", ""),
   DISABLED_AUTO_LINKING("103", ""),
@@ -19,7 +20,9 @@ public enum ErrorCode {
   DUPLICATE_AUTHORITY_SOURCE_FILE_ID("111",
       "Authority Source File with the given 'id' already exists."),
   NOT_EXISTED_AUTHORITY_SOURCE_FILE("112",
-    "Authority Source File with the given 'id' does not exists.");
+    "Authority Source File with the given 'id' does not exists."),
+  VIOLATION_OF_RELATION_BETWEEN_AUTHORITY_ARCHIVE_AND_SOURCE_FILE("113",
+    "Cannot complete operation on the entity due to it's relation with Authority Archive/Authority.");
 
   @Getter
   private final String code;